requirements.txt
fastapi
uvicorn
google-generativeai
firebase-admin
pydantic
pytest
httpx
